survey: Separate data helpers from interview utility helpers
The `evolution-common/src/utils/helpers.ts` file contains helper
functions both for manipulating general survey structure (responses,
validations, paths, etc) and getting data from the data model (persons,
trips, visited places, etc).

The latter are moved to the `services/odSurvey/helpers.ts` file and
should eventually all be replaced by calls to model objects, when we move
away from the current json structure to a more object oriented one.

Update all usages to use the new helper functions
